+++ packages/babel-cli/index.js
@@ -1,3 +1,1 @@
-throw new Error(
-  "Use the `@gerhobbelt/babel-core` package instead of `@gerhobbelt/babel-cli`."
-);
+throw new Error("Use the `@gerhobbelt/babel-core` package instead of `@gerhobbelt/babel-cli`.");