+++ packages/babel-template/src/parse.js
@@ -1,9 +1,6 @@
 // @flow
 import * as t from "@gerhobbelt/babel-types";
-import type {
-  TraversalAncestors,
-  TraversalHandler,
-} from "@gerhobbelt/babel-types";
+import type { TraversalAncestors, TraversalHandler } from "@gerhobbelt/babel-types";
 import { parse } from "@gerhobbelt/babel-parser";
 import { codeFrameColumns } from "@gerhobbelt/babel-code-frame";
 import type { TemplateOpts, ParserOpts } from "./options";